move version check to sdk_info.hpp and fix gcc issues (maybe..)
